+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This file documents major changes to the project in each version.
 ## Unreleased
 ### Added
   - Database tools for downloading game information: dl-covers, dl-data
+  - Database management tools: split-db
 
 ## v0.0.0 - 2018-05-09
 No features implemented yet.